+++ api/handlers.go
@@ -14,9 +14,11 @@ import (
 	"strings"
 	"time"
 
+	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/pkg/parsers/kernel"
 	versionpkg "github.com/docker/docker/pkg/version"
 	apitypes "github.com/docker/engine-api/types"
+	containertypes "github.com/docker/engine-api/types/container"
 	dockerfilters "github.com/docker/engine-api/types/filters"
 	"github.com/docker/swarm/cluster"
 	"github.com/docker/swarm/experimental"
@@ -71,9 +73,13 @@ func getInfo(c *context, w http.ResponseWriter, r *http.Request) {
 		info.SystemStatus = status
 	}
 
-	if kernelVersion, err := kernel.GetKernelVersion(); err == nil {
-		info.KernelVersion = kernelVersion.String()
+	kernelVersion := "<unknown>"
+	if kv, err := kernel.GetKernelVersion(); err != nil {
+		log.Warnf("Could not get kernel version: %v", err)
+	} else {
+		kernelVersion = kv.String()
 	}
+	info.KernelVersion = kernelVersion
 
 	for _, c := range c.cluster.Containers() {
 		info.Containers++
@@ -86,9 +92,13 @@ func getInfo(c *context, w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	if hostname, err := os.Hostname(); err == nil {
-		info.Name = hostname
+	hostname := "<unknown>"
+	if hn, err := os.Hostname(); err != nil {
+		log.Warnf("Could not get hostname: %v", err)
+	} else {
+		hostname = hn
 	}
+	info.Name = hostname
 
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(info)
@@ -107,9 +117,13 @@ func getVersion(c *context, w http.ResponseWriter, r *http.Request) {
 		BuildTime:    version.BUILDTIME,
 	}
 
-	if kernelVersion, err := kernel.GetKernelVersion(); err == nil {
-		version.KernelVersion = kernelVersion.String()
+	kernelVersion := "<unknown>"
+	if kv, err := kernel.GetKernelVersion(); err != nil {
+		log.Warnf("Could not get kernel version: %v", err)
+	} else {
+		kernelVersion = kv.String()
 	}
+	version.KernelVersion = kernelVersion
 
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(version)
@@ -171,11 +185,13 @@ func getImagesJSON(c *context, w http.ResponseWriter, r *http.Request) {
 	accepteds := filters.Get("node")
 	// this struct helps grouping images
 	// but still keeps their Engine infos as an array.
-	groupImages := make(map[string]dockerclient.Image)
+	groupImages := make(map[string]apitypes.Image)
 	opts := cluster.ImageFilterOptions{
-		All:        boolValue(r, "all"),
-		NameFilter: r.FormValue("filter"),
-		Filters:    filters,
+		apitypes.ImageListOptions{
+			All:       boolValue(r, "all"),
+			MatchName: r.FormValue("filter"),
+			Filters:   filters,
+		},
 	}
 	for _, image := range c.cluster.Images().Filter(opts) {
 		if len(accepteds) != 0 {
@@ -192,16 +208,16 @@ func getImagesJSON(c *context, w http.ResponseWriter, r *http.Request) {
 		}
 
 		// grouping images by Id, and concat their RepoTags
-		if entry, existed := groupImages[image.Id]; existed {
+		if entry, existed := groupImages[image.ID]; existed {
 			entry.RepoTags = append(entry.RepoTags, image.RepoTags...)
 			entry.RepoDigests = append(entry.RepoDigests, image.RepoDigests...)
-			groupImages[image.Id] = entry
+			groupImages[image.ID] = entry
 		} else {
-			groupImages[image.Id] = image.Image
+			groupImages[image.ID] = image.Image
 		}
 	}
 
-	images := []dockerclient.Image{}
+	images := []apitypes.Image{}
 
 	for _, image := range groupImages {
 		// de-duplicate RepoTags
@@ -254,7 +270,7 @@ func getNetworks(c *context, w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	out := []*dockerclient.NetworkResource{}
+	out := []*apitypes.NetworkResource{}
 	networks := c.cluster.Networks().Filter(filters.Get("name"), filters.Get("id"), types)
 	for _, network := range networks {
 		tmp := (*network).NetworkResource
@@ -271,8 +287,11 @@ func getNetworks(c *context, w http.ResponseWriter, r *http.Request) {
 func getNetwork(c *context, w http.ResponseWriter, r *http.Request) {
 	var id = mux.Vars(r)["networkid"]
 	if network := c.cluster.Networks().Uniq().Get(id); network != nil {
+		// there could be duplicate container endpoints in network, need to remove redundant
+		// see https://github.com/docker/swarm/issues/1969
+		cleanNetwork := network.RemoveDuplicateEndpoints()
 		w.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(w).Encode(network)
+		json.NewEncoder(w).Encode(cleanNetwork)
 		return
 	}
 	httpError(w, fmt.Sprintf("No such network: %s", id), http.StatusNotFound)
@@ -291,7 +310,7 @@ func getVolume(c *context, w http.ResponseWriter, r *http.Request) {
 
 // GET /volumes
 func getVolumes(c *context, w http.ResponseWriter, r *http.Request) {
-	volumes := struct{ Volumes []*dockerclient.Volume }{}
+	volumes := struct{ Volumes []*apitypes.Volume }{}
 
 	for _, volume := range c.cluster.Volumes() {
 		tmp := (*volume).Volume
@@ -572,7 +591,7 @@ func deleteContainers(c *context, w http.ResponseWriter, r *http.Request) {
 
 // POST /networks/create
 func postNetworksCreate(c *context, w http.ResponseWriter, r *http.Request) {
-	var request dockerclient.NetworkCreate
+	var request apitypes.NetworkCreate
 
 	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
 		httpError(w, err.Error(), http.StatusBadRequest)
@@ -590,12 +609,13 @@ func postNetworksCreate(c *context, w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
 	json.NewEncoder(w).Encode(response)
 }
 
 // POST /volumes/create
 func postVolumesCreate(c *context, w http.ResponseWriter, r *http.Request) {
-	var request dockerclient.VolumeCreateRequest
+	var request apitypes.VolumeCreateRequest
 
 	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
 		httpError(w, err.Error(), http.StatusBadRequest)
@@ -609,6 +629,7 @@ func postVolumesCreate(c *context, w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
 	json.NewEncoder(w).Encode(volume)
 }
 
@@ -685,7 +706,7 @@ func postImagesCreate(c *context, w http.ResponseWriter, r *http.Request) {
 // POST /images/load
 func postImagesLoad(c *context, w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusCreated)
+	w.WriteHeader(http.StatusOK)
 
 	// call cluster to load image on every node
 	wf := NewWriteFlusher(w)
@@ -749,7 +770,27 @@ func postContainersStart(c *context, w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := c.cluster.StartContainer(container); err != nil {
+	hostConfig := &dockerclient.HostConfig{
+		MemorySwappiness: -1,
+	}
+
+	buf, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		httpError(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+	r.Body.Close()
+
+	if len(buf) <= 2 {
+		hostConfig = nil
+	} else {
+		if err := json.Unmarshal(buf, hostConfig); err != nil {
+			httpError(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+	}
+
+	if err := c.cluster.StartContainer(container, hostConfig); err != nil {
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
@@ -910,7 +951,7 @@ func proxyNetworkDisconnect(c *context, w http.ResponseWriter, r *http.Request) 
 	r.Body = ioutil.NopCloser(bytes.NewBuffer(buf))
 
 	// Extract container info from r.Body copy
-	var disconnect dockerclient.NetworkDisconnect
+	var disconnect apitypes.NetworkDisconnect
 	if err := json.NewDecoder(bodyCopy).Decode(&disconnect); err != nil {
 		httpError(w, fmt.Sprintf("Container is not specified"), http.StatusNotFound)
 		return
@@ -966,7 +1007,7 @@ func proxyNetworkConnect(c *context, w http.ResponseWriter, r *http.Request) {
 	r.Body = ioutil.NopCloser(bytes.NewBuffer(buf))
 
 	// Extract container info from r.Body copy
-	var connect dockerclient.NetworkConnect
+	var connect apitypes.NetworkConnect
 	if err := json.NewDecoder(bodyCopy).Decode(&connect); err != nil {
 		httpError(w, fmt.Sprintf("Container is not specified"), http.StatusNotFound)
 		return
@@ -1114,6 +1155,7 @@ func postTagImage(c *context, w http.ResponseWriter, r *http.Request) {
 			httpError(w, err.Error(), http.StatusInternalServerError)
 		}
 	}
+	w.WriteHeader(http.StatusCreated)
 }
 
 // Proxy a request to a random node
@@ -1182,25 +1224,26 @@ func postBuild(c *context, w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	buildImage := &dockerclient.BuildImage{
-		DockerfileName: r.Form.Get("dockerfile"),
-		RepoName:       r.Form.Get("t"),
-		RemoteURL:      r.Form.Get("remote"),
+	buildImage := &apitypes.ImageBuildOptions{
+		Dockerfile:     r.Form.Get("dockerfile"),
+		Tags:           r.Form["t"],
+		RemoteContext:  r.Form.Get("remote"),
 		NoCache:        boolValue(r, "nocache"),
-		Pull:           boolValue(r, "pull"),
+		PullParent:     boolValue(r, "pull"),
 		Remove:         boolValue(r, "rm"),
 		ForceRemove:    boolValue(r, "forcerm"),
 		SuppressOutput: boolValue(r, "q"),
+		Isolation:      containertypes.Isolation(r.Form.Get("isolation")),
 		Memory:         int64ValueOrZero(r, "memory"),
 		MemorySwap:     int64ValueOrZero(r, "memswap"),
-		CpuShares:      int64ValueOrZero(r, "cpushares"),
-		CpuPeriod:      int64ValueOrZero(r, "cpuperiod"),
-		CpuQuota:       int64ValueOrZero(r, "cpuquota"),
-		CpuSetCpus:     r.Form.Get("cpusetcpus"),
-		CpuSetMems:     r.Form.Get("cpusetmems"),
+		CPUShares:      int64ValueOrZero(r, "cpushares"),
+		CPUPeriod:      int64ValueOrZero(r, "cpuperiod"),
+		CPUQuota:       int64ValueOrZero(r, "cpuquota"),
+		CPUSetCPUs:     r.Form.Get("cpusetcpus"),
+		CPUSetMems:     r.Form.Get("cpusetmems"),
 		CgroupParent:   r.Form.Get("cgroupparent"),
+		ShmSize:        int64ValueOrZero(r, "shmsize"),
 		Context:        r.Body,
-		BuildArgs:      make(map[string]string),
 	}
 
 	buildArgsJSON := r.Form.Get("buildargs")
@@ -1208,11 +1251,21 @@ func postBuild(c *context, w http.ResponseWriter, r *http.Request) {
 		json.Unmarshal([]byte(buildArgsJSON), &buildImage.BuildArgs)
 	}
 
-	authEncoded := r.Header.Get("X-Registry-Auth")
+	ulimitsJSON := r.Form.Get("ulimits")
+	if ulimitsJSON != "" {
+		json.Unmarshal([]byte(ulimitsJSON), &buildImage.Ulimits)
+	}
+
+	labelsJSON := r.Form.Get("labels")
+	if labelsJSON != "" {
+		json.Unmarshal([]byte(labelsJSON), &buildImage.Labels)
+	}
+
+	authEncoded := r.Header.Get("X-Registry-Config")
 	if authEncoded != "" {
-		buf, err := base64.URLEncoding.DecodeString(r.Header.Get("X-Registry-Auth"))
+		buf, err := base64.URLEncoding.DecodeString(r.Header.Get("X-Registry-Config"))
 		if err == nil {
-			json.Unmarshal(buf, &buildImage.Config)
+			json.Unmarshal(buf, &buildImage.AuthConfigs)
 		}
 	}
 
@@ -1248,6 +1301,7 @@ func postRenameContainer(c *context, w http.ResponseWriter, r *http.Request) {
 			httpError(w, err.Error(), http.StatusInternalServerError)
 		}
 	}
+	w.WriteHeader(http.StatusNoContent)
 
 }