+++ cluster/engine_test.go
@@ -7,6 +7,8 @@ import (
 	"testing"
 	"time"
 
+	"golang.org/x/net/context"
+
 	engineapi "github.com/docker/engine-api/client"
 	"github.com/docker/engine-api/types"
 	engineapimock "github.com/docker/swarm/api/mockclient"
@@ -115,7 +117,7 @@ func TestEngineConnectionFailure(t *testing.T) {
 	// Always fail.
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
-	apiClient.On("Info").Return(types.Info{}, errors.New("fail"))
+	apiClient.On("Info", mock.Anything).Return(types.Info{}, errors.New("fail"))
 
 	// Connect() should fail
 	assert.Error(t, engine.ConnectWithClient(client, apiClient))
@@ -134,7 +136,7 @@ func TestOutdatedEngine(t *testing.T) {
 	engine := NewEngine("test", 0, engOpts)
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
-	apiClient.On("Info").Return(types.Info{}, nil)
+	apiClient.On("Info", mock.Anything).Return(types.Info{}, nil)
 
 	assert.Error(t, engine.ConnectWithClient(client, apiClient))
 
@@ -154,12 +156,16 @@ func TestEngineCpusMemory(t *testing.T) {
 
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{}, nil)
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil)
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 
 	assert.NoError(t, engine.ConnectWithClient(client, apiClient))
@@ -180,12 +186,16 @@ func TestEngineSpecs(t *testing.T) {
 
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{}, nil)
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil)
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 
 	assert.NoError(t, engine.ConnectWithClient(client, apiClient))
@@ -211,15 +221,19 @@ func TestEngineState(t *testing.T) {
 
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 
 	// The client will return one container at first, then a second one will appear.
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{{Id: "one"}}, nil).Once()
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil).Once()
 	client.On("InspectContainer", "one").Return(&dockerclient.ContainerInfo{Config: &dockerclient.ContainerConfig{CpuShares: 100}}, nil).Once()
 	client.On("ListContainers", true, false, fmt.Sprintf("{%q:[%q]}", "id", "two")).Return([]dockerclient.Container{{Id: "two"}}, nil).Once()
 	client.On("InspectContainer", "two").Return(&dockerclient.ContainerInfo{Config: &dockerclient.ContainerConfig{CpuShares: 100}}, nil).Once()
@@ -263,13 +277,17 @@ func TestCreateContainer(t *testing.T) {
 	)
 
 	engine.setState(stateUnhealthy)
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{}, nil).Once()
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil).Once()
 
 	assert.NoError(t, engine.ConnectWithClient(client, apiClient))
 	assert.True(t, engine.isConnected())
@@ -284,9 +302,7 @@ func TestCreateContainer(t *testing.T) {
 	var auth *dockerclient.AuthConfig
 	client.On("CreateContainer", &mockConfig, name, auth).Return(id, nil).Once()
 	client.On("ListContainers", true, false, fmt.Sprintf(`{"id":[%q]}`, id)).Return([]dockerclient.Container{{Id: id}}, nil).Once()
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil).Once()
 	client.On("InspectContainer", id).Return(&dockerclient.ContainerInfo{Config: &config.ContainerConfig}, nil).Once()
 	container, err := engine.Create(config, name, false, auth)
 	assert.Nil(t, err)
@@ -309,9 +325,7 @@ func TestCreateContainer(t *testing.T) {
 	client.On("CreateContainer", &mockConfig, name, auth).Return("", dockerclient.ErrImageNotFound).Once()
 	client.On("CreateContainer", &mockConfig, name, auth).Return(id, nil).Once()
 	client.On("ListContainers", true, false, fmt.Sprintf(`{"id":[%q]}`, id)).Return([]dockerclient.Container{{Id: id}}, nil).Once()
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil).Once()
 	client.On("InspectContainer", id).Return(&dockerclient.ContainerInfo{Config: &config.ContainerConfig}, nil).Once()
 	container, err = engine.Create(config, name, true, auth)
 	assert.Nil(t, err)
@@ -323,9 +337,9 @@ func TestImages(t *testing.T) {
 	engine := NewEngine("test", 0, engOpts)
 	engine.setState(stateHealthy)
 	engine.images = []*Image{
-		{dockerclient.Image{Id: "a"}, engine},
-		{dockerclient.Image{Id: "b"}, engine},
-		{dockerclient.Image{Id: "c"}, engine},
+		{types.Image{ID: "a"}, engine},
+		{types.Image{ID: "b"}, engine},
+		{types.Image{ID: "c"}, engine},
 	}
 
 	result := engine.Images()
@@ -369,13 +383,17 @@ func TestUsedCpus(t *testing.T) {
 				mockInfo.NCPU = hn
 				cpuShares := int64(math.Ceil(float64(cn*1024) / float64(mockInfo.NCPU)))
 
-				apiClient.On("Info").Return(mockInfo, nil).Once()
-				apiClient.On("ServerVersion").Return(mockVersion, nil)
+				apiClient.On("Info", mock.Anything).Return(mockInfo, nil).Once()
+				apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+				apiClient.On("NetworkList", mock.Anything,
+					mock.AnythingOfType("NetworkListOptions"),
+				).Return([]types.NetworkResource{}, nil)
+				apiClient.On("VolumeList", mock.Anything,
+					mock.AnythingOfType("Args"),
+				).Return(types.VolumesListResponse{}, nil)
 				client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
-				client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
+				apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil).Once()
 				client.On("ListContainers", true, false, "").Return([]dockerclient.Container{{Id: "test"}}, nil).Once()
-				client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-				client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
 				client.On("InspectContainer", "test").Return(&dockerclient.ContainerInfo{Config: &dockerclient.ContainerConfig{CpuShares: cpuShares}}, nil).Once()
 
 				engine.ConnectWithClient(client, apiClient)
@@ -403,13 +421,17 @@ func TestContainerRemovedDuringRefresh(t *testing.T) {
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
 
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{container1, container2}, nil)
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
 	client.On("InspectContainer", "c1").Return(info1, errors.New("Not found"))
 	client.On("InspectContainer", "c2").Return(info2, nil)
 
@@ -431,16 +453,20 @@ func TestDisconnect(t *testing.T) {
 	client := mockclient.NewMockClient()
 	apiClient := engineapimock.NewMockClient()
 
-	apiClient.On("Info").Return(mockInfo, nil)
-	apiClient.On("ServerVersion").Return(mockVersion, nil)
+	apiClient.On("Info", mock.Anything).Return(mockInfo, nil)
+	apiClient.On("ServerVersion", mock.Anything).Return(mockVersion, nil)
+	apiClient.On("NetworkList", mock.Anything,
+		mock.AnythingOfType("NetworkListOptions"),
+	).Return([]types.NetworkResource{}, nil)
+	apiClient.On("VolumeList", mock.Anything,
+		mock.AnythingOfType("Args"),
+	).Return(types.VolumesListResponse{}, nil)
 	client.On("StartMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 	client.On("StopAllMonitorEvents", mock.Anything, mock.Anything, mock.Anything).Return()
 
 	// The client will return one container at first, then a second one will appear.
 	client.On("ListContainers", true, false, "").Return([]dockerclient.Container{{Id: "one"}}, nil).Once()
-	client.On("ListImages", mock.Anything).Return([]*dockerclient.Image{}, nil).Once()
-	client.On("ListVolumes", mock.Anything).Return([]*dockerclient.Volume{}, nil)
-	client.On("ListNetworks", mock.Anything).Return([]*dockerclient.NetworkResource{}, nil)
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil)
 	client.On("InspectContainer", "one").Return(&dockerclient.ContainerInfo{Config: &dockerclient.ContainerConfig{CpuShares: 100}}, nil).Once()
 	client.On("ListContainers", true, false, fmt.Sprintf("{%q:[%q]}", "id", "two")).Return([]dockerclient.Container{{Id: "two"}}, nil).Once()
 	client.On("InspectContainer", "two").Return(&dockerclient.ContainerInfo{Config: &dockerclient.ContainerConfig{CpuShares: 100}}, nil).Once()
@@ -459,4 +485,26 @@ func TestDisconnect(t *testing.T) {
 	assert.True(t, engine.state == stateDisconnected)
 	// Double disconnect shouldn't cause panic
 	engine.Disconnect()
+}
+
+func TestRemoveImage(t *testing.T) {
+	engine := NewEngine("test", 0, engOpts)
+
+	imageName := "test-image"
+	dIs := []types.ImageDelete{{Deleted: imageName}}
+
+	apiClient := engineapimock.NewMockClient()
+	apiClient.On("ImageList", mock.Anything, mock.AnythingOfType("ImageListOptions")).Return([]types.Image{}, nil)
+	apiClient.On("ImageRemove", context.TODO(),
+		mock.AnythingOfType("ImageRemoveOptions")).Return(dIs, nil)
+	engine.apiClient = apiClient
+
+	deletedImages, err := engine.RemoveImage("test-image", true)
+	if err != nil {
+		t.Errorf("encountered an unexpected error")
+	}
+	if deletedImages[0].Deleted != imageName {
+		t.Errorf("didn't get the image we removed")
+	}
+	apiClient.Mock.AssertExpectations(t)
 }